chore(storage): smallish refactoring series, part 1
Removed aliased type to bool.

Go aliased types (i.e. type A = B) are a syntactic sugar only.
This is different from type redefinitions (i.e. type A B), which create a new type for the compiler.

The original intent was to construct some kind of enumerated type, for a... boolean.

I chose to remove all this and just go back to plain bool.

* contributes #374

Signed-off-by: Frederic BIDON <frederic@oneconcern.com>