fix: detect and recover from corrupted blobs [HAZ-305]
On some rare but not impossible occasion, the written root key for a
file may become corrupted.

In that case, if the blob has been written, it will never be
overwritten again and it becomes impossible to recover that file,
even after several subsequent uploads.
What is worse, this situation goes completely undetected, and the corruption occurs silently until
we attempt to read the corrupted file.

We believe a similar situation may occur on any written blob.

This PR proposes a better handling of such situations. Even though it
remains impossible to totally guarantee against corruptions, what we can
do is:
* check back what's just been written when uploading and fail if we
detect a corruption
* overwrite outright corrupted blobs (e.g. empty, bad CRC) when
uploading

1. Every time we are about to write a root key blob (for each uploaded file), a
checksum check is carried out _before_ we decide to skip the write. If
the check is unsuccessful, this means that the existing key is
corrupted: a warning is emitted and the key is overwritten.

2. Every time we write a  blob (for each uploaded file), a
checksum check is carried out _after_ the write. The upload fails if
this is not successful (this check may be disabled with
--verify-hash=false).

3. Every time we are about to write a chunk blob, the "duplicate" check
only skips valid blobs. Corrupted blobs are overwritten.

4. Paranoid mode enabled with --verify-blob-hash (disabled by default).
Similar to what we do for root keys, with a slight performance penalty
(one more cloud store roundtrip to collect the blob size and CRC): this
checks back that the blob that has just been written is valid.

Signed-off-by: Frederic BIDON <frederic@oneconcern.com>